LayoutTests/http/tests/security/cross-frame-access-protocol-explicit-domain-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "https://127.0.0.1:8443".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "https". Protocols must match.

This test currently fails because we check the port and protocol even if document.domain is explicitly set (rdar://problem/5366437).


PASS: Cross frame access to https from http, after explicitly setting document.domain, was denied.


--------
Frame: 'aFrame'
--------
Inner iframe with explicit document.domain set.
